Taking a more structured approach to Chrome's various domains
So far modeling only the Runtime domain.

ChromeTabPage responds to #runtime, the message constructs new ChromeRuntime object which represents a domain available in dev tools. ChromeRuntime object encapsulates all of the arguments and calls, such as script evaluation.

Reifying a number of classes around all of that - ChromeMessage, ChromeResponse and ChromeResponseResult encapsulate communication with Chrome. Using Magritte-JSON for serializing and marshalling of data.

Messages sent to chrome that result in an exception will signala ChromeError.

Result of a message is generally returned as a dictionary. Domain objects can pass magritte descriptions for the objects they expect to be returned in the result. ChromeTabPage will then try to materialize results using those descriptions, via value-holding proxy ChromeResponseResult object.

Adding ChromeRemoteObject and other structs that are being referenced in Runtime domain, but not everything is there just yet.

Adding suppport for callbacks, using Promises. It is possible to #evaluate:callback: and #listenForEventsWith:callback: